feat(Messages): throttle user requests
