+++ .env_example
@@ -1,3 +1,7 @@
 BOT_TOKEN=
+
 #MAIN_CHANNEL_ID: TO WHERE THE GENERIC MESSAGES ARE SENT SUCH AS HELLO OR GOODBYE
 MAIN_CHANNEL_ID=
+
+#TIME IN SECONDS TO THROTTLE REQUESTS TO BOT
+MSG_THROTTLE_TIME=30
