+++ config/index.js
@@ -1,0 +1,14 @@
+const messages = {
+  enter: 'bom dia a todos menos a um',
+  exit: 'XAU AI _BORDAS_',
+};
+
+const activities = {
+  enter: 'Pepe as a Service',
+  exit: `Screw u guys I'm going home`,
+};
+
+module.exports = {
+  messages,
+  activities,
+};
