+++ package.json
@@ -11,7 +11,8 @@
   "license": "ISC",
   "dependencies": {
     "discord.js": "^11.4.2",
-    "dotenv": "^6.2.0"
+    "dotenv": "^6.2.0",
+    "lodash.throttle": "^4.1.1"
   },
   "devDependencies": {
     "eslint": "^5.14.1",
