+++ src/boot/index.js
