+++ src/events/index.js
@@ -1,0 +1,7 @@
+const ready = require('./ready');
+const message = require('./message');
+
+module.exports = {
+  ready,
+  message,
+};
